+++ azdev/operations/tests/__init__.py
@@ -25,6 +25,10 @@ from azdev.utilities import (
 
 logger = get_logger(__name__)
 
+_CORE_NAME_REGEX = re.compile(r'azure-cli-(?P<name>[^/\\]+)[/\\]azure[/\\]cli')
+_MOD_NAME_REGEX = re.compile(r'azure-cli[/\\]azure[/\\]cli[/\\]command_modules[/\\](?P<name>[^/\\]+)')
+_EXT_NAME_REGEX = re.compile(r'.*(?P<name>azext_[^/\\]+).*')
+
 
 def run_tests(tests, xml_path=None, discover=False, in_series=False,
               run_live=False, profile=None, last_failed=False, pytest_args=None):
@@ -103,13 +107,13 @@ def run_tests(tests, xml_path=None, discover=False, in_series=False,
 
 
 def _extract_module_name(path):
-    mod_name_regex = re.compile(r'azure-cli-([^/\\]+)[/\\]azure[/\\]cli')
-    ext_name_regex = re.compile(r'.*(azext_[^/\\]+).*')
 
-    try:
-        return re.search(mod_name_regex, path).group(1)
-    except AttributeError:
-        return re.search(ext_name_regex, path).group(1)
+    for expression in [_MOD_NAME_REGEX, _CORE_NAME_REGEX, _EXT_NAME_REGEX]:
+        match = re.search(expression, path)
+        if not match:
+            continue
+        return match.groupdict().get('name')
+    raise CLIError('unexpected error: unable to extract name from path: {}'.format(path))
 
 
 def _get_profile(profile):
@@ -203,6 +207,7 @@ def _discover_tests(profile):
         filepath = mod_path
         for comp in mod_name.split('-'):
             filepath = os.path.join(filepath, comp)
+
         mod_data = {
             'alt_name': 'main' if mod_name == 'azure-cli' else mod_name.replace(COMMAND_MODULE_PREFIX, ''),
             'filepath': os.path.join(filepath, 'tests'),
@@ -215,12 +220,13 @@ def _discover_tests(profile):
 
     logger.info('\nCommand Modules: %s', ', '.join([name for name, _ in command_modules]))
     for mod_name, mod_path in command_modules:
-        short_name = mod_name.replace(COMMAND_MODULE_PREFIX, '')
         mod_data = {
-            'alt_name': short_name,
+            # Modules don't technically have azure-cli-foo moniker anymore, but preserving
+            # for consistency.
+            'alt_name': '{}{}'.format(COMMAND_MODULE_PREFIX, mod_name),
             'filepath': os.path.join(
-                mod_path, 'azure', 'cli', 'command_modules', short_name, 'tests', profile_namespace),
-            'base_path': 'azure.cli.command_modules.{}.tests.{}'.format(short_name, profile_namespace),
+                mod_path, 'tests', profile_namespace),
+            'base_path': 'azure.cli.command_modules.{}.tests.{}'.format(mod_name, profile_namespace),
             'files': {}
         }
         tests = _discover_module_tests(mod_name, mod_data)